Use the geode nightly snapshots for the develop branch
In order to write examples for the current code, the geode-examples
develop branch should point to the geode nightly snapshots.

master should remain pointed to the latest release, and that is what users
will get when they clone.

Also, changing the download to use the version of geode from maven.